fix/CodeNumbers - inputs repaired
