+++ src/ui-lib/Checkbox/Checkbox.scss
@@ -12,6 +12,6 @@
     height: 24px;
     width: 24px;
     border: none;
-    background: url(../../ui-lib/checkmark.svg);
+    background: url(../checkmark.svg);
     background-color: #029992;
 }
